WIP: Run collection writes asynchronously on server.
The basic idea is that LocalCollection and _Mongo objects now have write methods
that take optional callbacks, and they run asynchronously and call the callback
when done. (For LocalCollection, nothing is actually asynchronous but just
appears to be.) The write methods on Collection wrap these write methods using
Meteor._wrapAsync.

Tests are passing, but needs to be cleaned up.